SWIK-147 moving thumbanil component to common components. Updated imports in DeckView and DeckCard
